Fix all clippy warnings and format code
- Collapse nested if-let chains in config.rs, mixer.rs, layout.rs
- Remove unnecessary u64-to-u64 casts in metadata.rs, streaming.rs
- Fix overindented doc list item in katla_audio/lib.rs
- Convert needless range loops to iter().enumerate() in mixer.rs
- Rename ScheduledEvent variants (remove At postfix)
- Introduce DrawInfo struct to reduce Widget::draw() from 9 to 7 params
- Introduce TreeViewCallbacks struct to reduce TreeView::new() from 8 to 7
- Run cargo fmt for consistent formatting

+++ katla_audio/src/metadata.rs
@@ -241,7 +241,7 @@ fn metadata_flac(path: &Path) -> Result<AudioMetadata, AudioError> {
     let channels = info.channels as u16;
     let sample_count = info
         .samples
-        .map(|s| s as u64 * channels as u64)
+        .map(|s| s * channels as u64)
         .unwrap_or(0);
     let duration_secs = if sample_rate > 0 {
         info.samples
